chore(bendpoints): extract bendpoint preview to separate file
