+++ lib/features/bendpoints/index.js
@@ -3,6 +3,7 @@ import RulesModule from '../rules';
 
 import Bendpoints from './Bendpoints';
 import BendpointMove from './BendpointMove';
+import BendpointMovePreview from './BendpointMovePreview';
 import ConnectionSegmentMove from './ConnectionSegmentMove';
 import BendpointSnapping from './BendpointSnapping';
 
@@ -12,9 +13,10 @@ export default {
     DraggingModule,
     RulesModule
   ],
-  __init__: [ 'bendpoints', 'bendpointSnapping' ],
+  __init__: [ 'bendpoints', 'bendpointSnapping', 'bendpointMovePreview' ],
   bendpoints: [ 'type', Bendpoints ],
   bendpointMove: [ 'type', BendpointMove ],
+  bendpointMovePreview: [ 'type', BendpointMovePreview ],
   connectionSegmentMove: [ 'type', ConnectionSegmentMove ],
   bendpointSnapping: [ 'type', BendpointSnapping ]
 };
